+++ nengo_edge/saved_model_runner.py
@@ -2,7 +2,7 @@
 
 import warnings
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import numpy as np
 
@@ -10,9 +10,10 @@ from nengo_edge import config
 
 try:
     import tensorflow as tf
-    from tensorflow_text import FastSentencepieceTokenizer
+    from tensorflow_text import SentencepieceTokenizer
 
     from nengo_edge import ragged  # pylint: disable=ungrouped-imports
+
 except ImportError:  # pragma: no cover
     warnings.warn("TensorFlow is not installed; cannot use SavedModelRunner.")
 
@@ -30,20 +31,18 @@ class SavedModelRunner:
         )
 
         if self.model_params["type"] == "asr":
-            if (
-                "tokenizer_file" in self.postprocessing
-                and (self.directory / self.postprocessing["tokenizer_file"]).exists()
-            ):
-                self.tokenizer = FastSentencepieceTokenizer(
-                    (
-                        self.directory / self.postprocessing["tokenizer_file"]
-                    ).read_bytes()
-                )
-            else:
-                self.tokenizer = None
+            self.tokenizer = self.get_tokenizer(self.postprocessing)
+            if self.tokenizer is None:
                 warnings.warn(
                     "No tokenizer model found, cannot decode ASR outputs. "
                     "Consider re-downloading ASR run artifacts."
+                )
+        elif self.model_params["type"] == "nlp":
+            self.tokenizer = self.get_tokenizer(self.preprocessing)
+            if self.tokenizer is None:
+                raise ValueError(
+                    "No tokenizer model found and is required to process string inputs."
+                    "Please re-download the NLP artifacts."
                 )
 
         self.reset_state()
@@ -52,6 +51,34 @@ class SavedModelRunner:
         """Reset the internal state of the model to initial conditions."""
 
         self.state: Optional[List[tf.Tensor]] = None
+
+    def get_tokenizer(
+        self, params: Dict[str, str]
+    ) -> Optional["SentencepieceTokenizer"]:
+        """
+        Load a ``tensorflow_text.SentencepieceTokenizer``.
+
+        Returns None if no ``tokenizer_file`` exists in the load directory.
+
+        Parameters
+        ----------
+        params: Dict[str, str]
+            Configuration dictionary containing the name of the
+            ``tokenizer_file`` found in the save directory.
+
+        Returns
+        -------
+        tokenizer: ``tensorflow_text.SentencepieceTokenizer``
+            A sentencepiece tokenizer with a trained vocabulary.
+        """
+        if (
+            "tokenizer_file" in params
+            and (self.directory / params["tokenizer_file"]).exists()
+        ):
+            return SentencepieceTokenizer(
+                (self.directory / params["tokenizer_file"]).read_bytes()
+            )
+        return None
 
     def _run_model(self, inputs: np.ndarray) -> np.ndarray:
         """
@@ -98,27 +125,40 @@ class SavedModelRunner:
 
         return outputs[0].numpy()
 
-    def run(self, inputs: np.ndarray) -> np.ndarray:
+    def run(self, inputs: Union[np.ndarray, List[str]]) -> np.ndarray:
         """
         Run the model on the given inputs.
 
-        Also applies model type specific actions to outputs. For example,
-        ASR model outputs are automatically detokenized to construct
-        a string prediction.
+        This function applies model specific actions to inputs or outputs.
+        For example, ASR model outputs are automatically detokenized to construct
+        a string prediction. For NLP models, the list of strings is
+        automatically processed into integer tokens from a saved
+        vocabulary.
 
         Parameters
         ----------
-        inputs : np.ndarray
-            Model input values (should have shape ``(batch_size, input_steps)``).
+        inputs : Union[np.ndarray, List[str]]
+            For audio based models, such as KWS and ASR, model input values
+            should be stored in a numpy array with a shape of
+            ``(batch_size, input_steps)``.
+            For text based NLP models, input strings should be stored in a list with
+            a length equal to ``batch_size``.
 
         Returns
         -------
         outputs : np.ndarray
             Model output values with shape ``(batch_size, output_steps, output_d)``
-            for KWS and ``(batch_size)`` for ASR.
+            for KWS, ``(batch_size)`` for ASR and ``(batch_size, output_d)`` for NLP.
         """
 
-        if inputs.dtype == object and self.model_params["type"] == "kws":
+        if isinstance(inputs[0], str):
+            # Process string inputs
+            assert self.model_params["type"] == "nlp"
+            assert self.tokenizer is not None
+            inputs = self.tokenizer.tokenize([s.lower() for s in inputs]).numpy()
+
+        assert isinstance(inputs, np.ndarray)
+        if self.model_params["type"] == "kws" and inputs.dtype == object:
             raise NotImplementedError("KWS models do not support ragged inputs")
 
         outputs = self._run_model(inputs)